refactored cid dag metadata in examples
