+++ examples/common.js
@@ -1,10 +1,5 @@
-import { blake2AsU8a } from '@polkadot/util-crypto'
-import * as multihash from 'multiformats/hashes/digest'
-import { CID } from 'multiformats/cid'
 import { Keyring } from '@polkadot/keyring';
 import { getPolkadotSigner } from '@polkadot-api/signer';
-import * as dagPB from '@ipld/dag-pb'
-import { UnixFS } from 'ipfs-unixfs'
 import { createCanvas } from "canvas";
 import fs from "fs";
 import assert from "assert";
@@ -13,35 +8,6 @@ export async function waitForNewBlock() {
     // TODO: wait for a new block.
     console.log('🛰 Waiting for new block...')
     return new Promise(resolve => setTimeout(resolve, 7000))
-}
-
-/**
- * Create CID for data.
- * Default to `0x55 (raw)` with blake2b_256 hash.
- *
- * 0xb220:
- * - 0xb2 = the multihash algorithm family for BLAKE2b
- * - 0x20 = the digest length in bytes (32 bytes = 256 bits)
- *
- * See: https://github.com/multiformats/multicodec/blob/master/table.csv
- */
-export async function cidFromBytes(bytes, cidCodec = 0x55, mhCode = 0xb220) {
-    console.log(`[CID]: Using cidCodec: ${cidCodec} and mhCode: ${mhCode}`);
-    let mh;
-    switch (mhCode) {
-        case 0xb220: // blake2b-256
-            mh = multihash.create(mhCode, blake2AsU8a(bytes));
-            break;
-
-        default:
-            throw new Error("Unhandled multihash code: " + mhCode)
-    }
-    return CID.createV1(cidCodec, mh)
-}
-
-export function convertCid(cid, cidCodec) {
-    const mh = cid.multihash;
-    return CID.createV1(cidCodec, mh);
 }
 
 /**
@@ -68,50 +34,6 @@ export function setupKeyringAndSigners(sudoSeed, accountSeed) {
         whoSigner,
         whoAddress: whoAccount.address
     };
-}
-
-/**
- * Build a UnixFS DAG-PB file node from raw chunks.
- *
- * (By default with SHA2 multihash)
- */
-export async function buildUnixFSDagPB(chunks, mhCode = 0x12) {
-    if (!chunks?.length) {
-        throw new Error('❌ buildUnixFSDag: chunks[] is empty')
-    }
-
-    // UnixFS blockSizes = sizes of child blocks
-    const blockSizes = chunks.map(c => c.len)
-
-    console.log(`🧩 Building UnixFS DAG from chunks:
-  • totalChunks: ${chunks.length}
-  • blockSizes: ${blockSizes.join(', ')}`)
-
-    // Build UnixFS file metadata (no inline data here)
-    const fileData = new UnixFS({
-        type: 'file',
-        blockSizes
-    })
-
-    // DAG-PB node: our file with chunk links
-    const dagNode = dagPB.prepare({
-        Data: fileData.marshal(),
-        Links: chunks.map(c => ({
-            Name: '',
-            Tsize: c.len,
-            Hash: c.cid
-        }))
-    })
-
-    // Encode DAG-PB
-    const dagBytes = dagPB.encode(dagNode)
-
-    // Hash DAG to produce CIDv1
-    const rootCid = await cidFromBytes(dagBytes, dagPB.code, mhCode)
-
-    console.log(`✅ DAG root CID: ${rootCid.toString()}`)
-
-    return { rootCid, dagBytes }
 }
 
 /**
@@ -171,5 +93,28 @@ export function filesAreEqual(path1, path2) {
 
     for (let i = 0; i < data1.length; i++) {
         assert.deepStrictEqual(data1[i], data2[i])
+    }
+}
+
+export async function fileToDisk(outputPath, fullBuffer) {
+    await new Promise((resolve, reject) => {
+        const ws = fs.createWriteStream(outputPath);
+        ws.write(fullBuffer);
+        ws.end();
+        ws.on('finish', resolve);
+        ws.on('error', reject);
+    });
+    console.log(`💾 File saved to: ${outputPath}`);
+}
+
+export class NonceManager {
+    constructor(initialNonce) {
+        this.nonce = initialNonce; // BN instance from api.query.system.account
+    }
+
+    getAndIncrement() {
+        const current = this.nonce;
+        this.nonce = this.nonce.addn(1); // increment BN
+        return current;
     }
 }
